bevy_rapier2d
Module dynamics

Components related to physics dynamics (rigid-bodies, velocities, etc.)

Structs

AdditionalSolverIterations

Set the additional number of solver iterations run for a rigid-body and everything interacting with it.

Ccd

Information used for Continuous-Collision-Detection.

Damping

Damping factors to gradually slow down a RigidBody.

Dominance

The dominance groups of a RigidBody.

ExternalForce

Constant external forces applied continuously to a RigidBody.

ExternalImpulse

Instantaneous external impulse applied continuously to a RigidBody.

FixedJoint

A fixed joint, locks all relative motion between two bodies.

FixedJointBuilder

Create fixed joints using the builder pattern.

GenericJoint

The description of any joint.

GenericJointBuilder

Create generic joints using the builder pattern.

GravityScale

Gravity is multiplied by this scaling factor before it’s applied to this RigidBody.

ImpulseJoint

An impulse-based joint attached to two entities.

JointAxesMask

A bit mask identifying multiple degrees of freedom of a joint.

LockedAxes

Flags affecting the behavior of the constraints solver for a given contact manifold.

MassModifiedEvent

Entity that likely had their mass properties changed this frame.

MassProperties

Center-of-mass, mass, and angular inertia.

MultibodyJoint

An joint based on generalized coordinates, attached to two entities.

PrismaticJoint

A prismatic joint, locks all relative motion between two bodies except for translation along the joint’s principal axis.

PrismaticJointBuilder

Create prismatic joints using the builder pattern.

RapierImpulseJointHandle

The handle of an impulse joint added to the physics scene.

RapierMultibodyJointHandle

The handle of a multibody joint added to the physics scene.

RapierRigidBodyHandle

The Rapier handle of a RigidBody that was inserted to the physics scene.

ReadMassProperties

Center-of-mass, mass, and angular inertia.

RevoluteJoint

A revolute joint, locks all relative motion except for rotation along the joint’s principal axis.

RevoluteJointBuilder

Create revolute joints using the builder pattern.

RigidBodyDisabled

Indicates whether or not the RigidBody is disabled explicitly by the user.

RopeJoint

A rope joint, limits the maximum distance between two bodies

RopeJointBuilder

Create rope joints using the builder pattern.

Sleeping

The activation status of a body.

SoftCcd

Sets the maximum prediction distance Soft Continuous Collision-Detection.

SpringJoint

A spring-damper joint, applies a force proportional to the distance between two objects.

SpringJointBuilder

A SpringJoint joint using the builder pattern.

TransformInterpolation

If the TimestepMode::Interpolated mode is set and this component is present, the associated RigidBody will have its position automatically interpolated between the last two RigidBody positions set by the physics engine.

Velocity

The velocity of a RigidBody.

Enums

AdditionalMassProperties

Mass-properties of a RigidBody, added to the contributions of its attached colliders.

CoefficientCombineRule

Rules used to combine two coefficients.

JointAxis

Identifiers of degrees of freedoms of a joint.

MotorModel

The spring-like model used for constraints resolution.

RigidBody

A RigidBody.

TypedJoint

Wrapper enum over a specific joint.