bevy_rapier2d
Module geometry

Components related to physics geometry (colliders, collision-groups, etc.)

Re-exports

pub use self::shape_views::ColliderView;

Modules

shape_views

Wrappers around Rapier shapes to access their properties.

Structs

ActiveCollisionTypes

Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

ActiveEvents

Flags affecting the events generated for this Collider.

ActiveHooks

Flags affecting the behavior of the constraints solver for a given contact manifold.

Collider

A geometric entity that can be attached to a RigidBody so it can be affected by contacts and intersection queries.

ColliderDisabled

Indicates whether or not the collider is disabled explicitly by the user.

CollidingEntities

Component which will be filled (if present) with a list of entities with which the current entity is currently in contact.

CollisionGroups

Pairwise collision filtering using bit masks.

ContactForceEventThreshold

The total force magnitude beyond which a ContactForceEvent can be emitted.

ContactSkin

Sets the contact skin of the collider.

Friction

The friction affecting a Collider.

Group

A bit mask identifying groups for interaction.

PointProjection

Result of the projection of a point on a shape.

RapierColliderHandle

The Rapier handle of a collider that was inserted to the physics scene.

RayIntersection

Structure containing the result of a successful ray cast.

Restitution

The restitution affecting a Collider.

Sensor

Indicates whether or not the Collider is a sensor.

ShapeCastHit

The result of a shape cast.

ShapeCastHitDetails

In depth information about a shape-cast hit.

ShapeCastOptions

Configuration for controlling the behavior of time-of-impact (i.e. shape-casting) calculations.

SolverFlags

Flags affecting the behavior of the constraints solver for a given contact manifold.

SolverGroups

Pairwise constraints resolution filtering using bit masks.

TriMeshFlags

The status of the cell of an heightfield.

VHACDParameters

Parameters controlling the VHACD convex decomposition.

Enums

ColliderMassProperties

Custom mass-properties of a Collider.

ColliderScale

Overwrites the default application of GlobalTransform scale to a Collider’s shapes.

FillMode

Controls how the voxelization determines which voxel needs to be considered empty, and which ones will be considered full.

ShapeCastStatus

The status of the time-of-impact computation algorithm.

Functions

get_snapped_scale

We restrict the scaling increment to 1.0e-4, to avoid numerical jitter due to the extraction of scaling factor from the GlobalTransform matrix.