bevy_rapier2d::geometry
Module shape_views

Wrappers around Rapier shapes to access their properties.

Structs

BallView

Read-only access to the properties of a ball.

BallViewMut

Read-write access to the properties of a ball.

CapsuleView

Read-only access to the properties of a capsule.

CapsuleViewMut

Read-write access to the properties of a capsule.

CompoundView

Read-only access to the properties of a compound shape.

ConvexPolygonView

Read-only access to the properties of a convex polygon.

ConvexPolygonViewMut

Read-write access to the properties of a convex polygon.

CuboidView

Read-only access to the properties of a cuboid.

CuboidViewMut

Read-write access to the properties of a cuboid.

HalfSpaceView

Read-only access to the properties of a half-space.

HalfSpaceViewMut

Read-write access to the properties of a half-space.

HeightFieldView

Read-only access to the properties of a heightfield.

HeightFieldViewMut

Read-write access to the properties of a heightfield.

PolylineView

Read-only access to the properties of a polyline.

PolylineViewMut

Read-write access to the properties of a polyline.

RoundConvexPolygonView

Read-only access to the properties of a round shape.

RoundConvexPolygonViewMut

Read-write access to the properties of a round shape.

RoundCuboidView

Read-only access to the properties of a round shape.

RoundCuboidViewMut

Read-write access to the properties of a round shape.

RoundTriangleView

Read-only access to the properties of a round shape.

RoundTriangleViewMut

Read-write access to the properties of a round shape.

SegmentView

Read-only access to the properties of a segment.

SegmentViewMut

Read-write access to the properties of a segment.

TriMeshView

Read-only access to the properties of a triangle mesh.

TriMeshViewMut

Read-write access to the properties of a triangle mesh.

TriangleView

Read-only access to the properties of a triangle.

TriangleViewMut

Read-write access to the properties of a triangle.

Enums

ColliderView

Read-only access to the properties of a collider.

Type Aliases

HeightFieldCellStatus

Indicates if a cell of a heightfield is removed or not. Set this to false for a removed cell.