Module bevy_rapier2d::physics

Plugins, resources, and components for physics simulation.

Structs

ColliderBundle

ColliderComponentsSet

ColliderHandleComponent
A component representing a collider that is being handled by a Rapier physics World.

JointBuilderComponent
Component responsible for initializing a Rapier joint.

JointHandleComponent
A component representing a joint added to the JointSet resource.

JointsEntityMap
HashMaps of Bevy Entity to Rapier handles

ModificationTracker

PhysicsHooksWithQueryObject

QueryComponentSetMut

QueryPipelineColliderComponentsSet

RapierConfiguration
A resource for specifying configuration information for the physics simulation

RapierPhysicsPlugin
A plugin responsible for setting up a full Rapier physics simulation pipeline and resources.

RigidBodyBundle

RigidBodyComponentsSet

RigidBodyHandleComponent
A component representing a rigid-body that is being handled by a Rapier physics World.

SimulationToRenderTime
Difference between simulation and rendering time

Enums

ColliderPositionSync

PhysicsStages
The names of the default App stages

RigidBodyPositionSync

TimestepMode
The different ways of adjusting the timestep length.

Constants

TRANSFORM_SYNC_STAGE
The stage where the physics transform are output to the Bevy Transform.

Traits

BundleBuilder

IntoEntity

IntoHandle

PhysicsHooksWithQuery

Functions

attach_bodies_and_colliders_system
System responsible for creating a Rapier rigid-body and collider from their builder resources.

collect_removals
System responsible for collecting the entities with removed rigid-bodies, colliders, or joints.

create_joints_system
System responsible for creating Rapier joints from their builder resources.

finalize_collider_attach_to_bodies
System responsible for creating a Rapier rigid-body and collider from their builder resources.

step_world_system
System responsible for performing one timestep of the physics world.

sync_transforms
System responsible for writing the rigid-bodies positions into the Bevy translation and rotation components.

Type Definitions

ColliderComponentsQuery

ComponentSetQueryMut

NoUserData

QueryPipelineColliderComponentsQuery

RigidBodyComponentsQuery