bevy_rapier2d
Module pipeline

Components and resources related to the physics simulation workflow (events, hooks, etc.)

Structs

ContactForceEvent

Event occurring when the sum of the magnitudes of the contact forces between two colliders exceed a threshold (ContactForceEventThreshold).

ContactModificationContextView

Read-write access to the properties of a contact modification context.

PairFilterContextView

Read-only access to the properties of a collision pair filter context.

QueryFilter

A filter that describes what collider should be included or excluded from a scene query.

QueryFilterFlags

Flags for excluding whole sets of colliders from a scene query.

Enums

CollisionEvent

Events occurring when two colliders start or stop colliding

Traits

BevyPhysicsHooks

User-defined functions called by the physics engines during one timestep in order to customize its behavior.