bevy_rapier2d::pipeline
Struct QueryFilter

pub struct QueryFilter<'a> {
    pub flags: QueryFilterFlags,
    pub groups: Option<CollisionGroups>,
    pub exclude_collider: Option<Entity>,
    pub exclude_rigid_body: Option<Entity>,
    pub predicate: Option<&'a dyn Fn(Entity) -> bool>,
}

A filter that describes what collider should be included or excluded from a scene query.

For testing manually check RapierContext::with_query_filter.

Fields
flags: QueryFilterFlags

Flags indicating what particular type of colliders should be excluded.

groups: Option<CollisionGroups>

If set, only colliders with collision groups compatible with this one will be included in the scene query.

exclude_collider: Option<Entity>

If set, the collider attached to that entity will be excluded from the query.

exclude_rigid_body: Option<Entity>

If set, any collider attached to the rigid-body attached to that entity will be excluded from the query.

predicate: Option<&'a dyn Fn(Entity) -> bool>

If set, any collider for which this closure returns false.

Implementations

impl<'a> QueryFilter<'a>

pub fn new() -> Self

A query filter that doesnโt exclude any collider.

pub fn exclude_fixed() -> Self

Exclude from the query any collider attached to a fixed rigid-body and colliders with no rigid-body attached.

pub fn exclude_kinematic() -> Self

Exclude from the query any collider attached to a kinematic rigid-body.

pub fn exclude_dynamic() -> Self

Exclude from the query any collider attached to a dynamic rigid-body.

pub fn only_dynamic() -> Self

Excludes all colliders not attached to a dynamic rigid-body.

pub fn only_kinematic() -> Self

Excludes all colliders not attached to a kinematic rigid-body.

pub fn only_fixed() -> Self

Exclude all colliders attached to a non-fixed rigid-body (this will not exclude colliders not attached to any rigid-body).

pub fn exclude_sensors(self) -> Self

Exclude from the query any collider that is a sensor.

pub fn exclude_solids(self) -> Self

Exclude from the query any collider that is not a sensor.

pub fn groups(self, groups: CollisionGroups) -> Self

Only colliders with collision groups compatible with this one will be included in the scene query.

pub fn exclude_collider(self, collider: Entity) -> Self

Set the collider that will be excluded from the scene query.

pub fn exclude_rigid_body(self, rigid_body: Entity) -> Self

Set the rigid-body that will be excluded from the scene query.

pub fn predicate(self, predicate: &'a impl Fn(Entity) -> bool) -> Self

Set the predicate to apply a custom collider filtering during the scene query.