bevy_rapier2d
Module plugin

The physics plugin and systems.

Modules

systems

Systems responsible for interfacing our Bevy components with the Rapier physics engine.

Structs

ContactManifoldView

Read-only access to the properties of a contact manifold.

ContactPairView

Read-only access to the properties of a contact pair.

ContactView

Read-only access to the properties of a single contact.

DefaultRapierContext

Marker component for to access the default RapierContext.

RapierConfiguration

A component for specifying configuration information for the physics simulation

RapierContext

The Rapier context, containing all the state of the physics engine.

RapierContextAccess

Utility SystemParam to easily access any RapierContext immutably

RapierContextEntityLink

This is a component applied to any entity containing a rapier handle component. The inner Entity referred to has the component RapierContext responsible for handling its rapier data.

RapierPhysicsPlugin

A plugin responsible for setting up a full Rapier physics simulation pipeline and resources.

RapierTransformPropagateSet

A set for rapier’s copy of Bevy’s transform propagation systems.

ReadDefaultRapierContext

Utility SystemParam to easily access the single default RapierContext immutably.

SimulationToRenderTime

Difference between simulation and rendering time

SolverContactView

Read-only access to the properties of a single solver contact.

WriteDefaultRapierContext

Utility SystemParam to easily access the single default RapierContext mutably.

WriteRapierContext

Utility SystemParam to easily access any RapierContext mutably

Enums

PhysicsSet

SystemSet for each phase of the plugin.

RapierContextInitialization

Specifies a default configuration for the default RapierContext

TimestepMode

The different ways of adjusting the timestep length each frame.

Type Aliases

NoUserData

No specific user-data is associated to the hooks.