bevy_rapier2d::plugin
Struct ReadDefaultRapierContext

pub struct ReadDefaultRapierContext<'w, 's, T: Component = DefaultRapierContext> { /* private fields */ }

Utility SystemParam to easily access the single default RapierContext immutably.

SAFETY: Dereferencing this struct will panic if its underlying query fails. See RapierContextAccess for a safer alternative.

Implementations

impl<'w, 's, T: Component> ReadDefaultRapierContext<'w, 's, T>

pub fn single(&self) -> &RapierContext

Use this method if you only have one RapierContext.

SAFETY: This method will panic if its underlying query fails. See RapierContextAccess for a safe alternative.