bevy_rapier2d::plugin
Module systems

Systems responsible for interfacing our Bevy components with the Rapier physics engine.

Functions

apply_collider_user_changes

System responsible for applying changes the user made to a collider-related component.

apply_initial_rigid_body_impulses

This applies the initial impulse given to a rigid-body when it is created.

apply_joint_user_changes

System responsible for applying changes the user made to a joint component.

apply_rigid_body_user_changes

System responsible for applying changes the user made to a rigid-body-related component.

apply_scale

System responsible for applying GlobalTransform scale and/or ColliderScale to colliders.

init_colliders

System responsible for creating new Rapier colliders from the related bevy_rapier components.

init_joints

System responsible for creating new Rapier joints from the related bevy_rapier components.

init_rigid_bodies

System responsible for creating new Rapier rigid-bodies from the related bevy_rapier components.

on_add_entity_with_parent

If an entity is turned into the child of something with a physics context link, the child should become a part of that physics context

on_change_context

reacts to modifications to RapierContextEntityLink to move an entity’s physics data from a context to another.

step_simulation

System responsible for advancing the physics simulation, and updating the internal state for scene queries.

sync_removals

System responsible for removing from Rapier the rigid-bodies/colliders/joints which had their related bevy_rapier components removed by the user (through component removal or despawn).

update_character_controls

System responsible for applying the character controller translation to the underlying collider.

update_colliding_entities

Adds entity to CollidingEntities on starting collision and removes from it when the collision ends.

writeback_mass_properties

System responsible for writing updated mass properties back into the ReadMassProperties component.

writeback_rigid_bodies

System responsible for writing the result of the last simulation step into our bevy_rapier components and the GlobalTransform component.

Type Aliases

ColliderComponents

Components related to colliders.

RigidBodyComponents

Components related to rigid-bodies.

RigidBodyWritebackComponents

Components that will be updated after a physics step.