bevy_rapier2d::plugin::systems
Function apply_initial_rigid_body_impulses

pub fn apply_initial_rigid_body_impulses(
    context: WriteRapierContext<'_, '_>,
    init_impulses: Query<'_, '_, (Entity, &RapierContextEntityLink, &mut ExternalImpulse), Without<RapierRigidBodyHandle>>,
)

This applies the initial impulse given to a rigid-body when it is created.

This cannot be done inside init_rigid_bodies because impulses require the rigid-body mass to be available, which it was not because colliders were not created yet. As a result, we run this system after the collider creation.