Module bevy_rapier2d::prelude

Re-exports

pub use super::physics::IntoEntity;

pub use super::physics::IntoHandle;

Modules

nalgebra
nalgebra

Macros

point
Construct a fixed-size point directly from data.

vector
Construct a fixed-size column vector directly from data.

Structs

ActiveCollisionTypes
Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

ActiveEvents
Flags affecting the events generated for this collider.

ActiveHooks
Flags affecting the behavior of the constraints solver for a given contact manifold.

BallJoint
A joint that removes all relative linear motion between a pair of points on two bodies.

BroadPhase
A broad-phase combining a Hierarchical Grid and Sweep-and-Prune.

CCDSolver
Solver responsible for performing motion-clamping on fast-moving bodies.

ChannelEventCollector
A physics event handler that collects events into a crossbeam channel.

ColliderBroadPhaseData
Data associated to a collider that takes part to a broad-phase algorithm.

ColliderBundle

ColliderChanges
Flags describing how the collider has been modified by the user.

ColliderComponentsSet

ColliderDebugRender
The desired render color of a Rapier collider.

ColliderFlags
A set of flags for controlling collision/intersection filtering, modification, and events.

ColliderHandle
The unique identifier of a collider added to a collider set.

ColliderMaterial
The constraints solver-related properties of this collider (friction, restitution, etc.)

ColliderParent
Information about the rigid-body this collider is attached to.

ColliderPosition
The position of a collider.

CollisionPipeline
The collision pipeline, responsible for performing collision detection between colliders.

Compound
A compound shape with an aabb bounding volume.

ContactData
A single contact between two collider.

ContactManifoldData
A contact manifold between two colliders.

ContactModificationContext
Context given to custom contact modifiers to modify the contacts seen by the constraints solver.

ContactPair
The description of all the contacts between a pair of colliders.

ConvexPolygon
A 2D convex polygon.

FixedJoint
A joint that prevents all relative movement between two bodies.

HalfSpace
A half-space delimited by an infinite plane.

IntegrationParameters
Parameters for a time-step of the physics engine.

InteractionGraph
A graph where nodes are collision objects and edges are contact or proximity algorithms.

InteractionGroups
Pairwise filtering using bit masks.

IntersectionEvent
Events occurring when two collision objects start or stop being in close proximity, contact, or disjoint.

IslandManager
Structure responsible for maintaining the set of active rigid-bodies, and putting non-moving rigid-bodies to sleep to save computation times.

Joint
A joint attached to two bodies.

JointBuilderComponent
Component responsible for initializing a Rapier joint.

JointHandle
The unique identifier of a joint added to the joint set. The unique identifier of a collider added to a collider set.

JointSet
A set of joints that can be handled by a physics World.

MassProperties
The local mass properties of a rigid-body.

NarrowPhase
The narrow-phase responsible for computing precise contact information between colliders.

PairFilterContext
Context given to custom collision filters to filter-out collisions.

PhysicsHooksWithQueryObject

PhysicsPipeline
The physics pipeline, responsible for stepping the whole physics simulation.

PolygonalFeature
A polygonal feature representing the local polygonal approximation of a vertex, or face, of a convex shape.

Polyline
A polyline.

PrismaticJoint
A joint that removes all relative motion between two bodies, except for the translations along one axis.

QueryPipeline
A pipeline for performing queries on all the colliders of a scene.

QueryPipelineColliderComponentsSet

RapierConfiguration
A resource for specifying configuration information for the physics simulation

RapierPhysicsPlugin
A plugin responsible for setting up a full Rapier physics simulation pipeline and resources.

RapierRenderPlugin
Plugin responsible for creating meshes to render the Rapier physics scene.

RigidBodyActivation
The rb_activation status of a body.

RigidBodyBundle

RigidBodyCcd
Information used for Continuous-Collision-Detection.

RigidBodyChanges
Flags describing how the rigid-body has been modified by the user.

RigidBodyColliders
The set of colliders attached to this rigid-bodies.

RigidBodyComponentsSet

RigidBodyDamping
Damping factors to progressively slow down a rigid-body.

RigidBodyDominance
The dominance groups of a rigid-body.

RigidBodyForces
The user-defined external forces applied to this rigid-body.

RigidBodyHandle
The unique handle of a rigid body added to a RigidBodySet.

RigidBodyIds
Internal identifiers used by the physics engine.

RigidBodyMassProps
The mass properties of this rigid-bodies.

RigidBodyMassPropsFlags
Flags affecting the behavior of the constraints solver for a given contact manifold.

RigidBodyPosition
The position of this rigid-body.

RigidBodyVelocity
The velocities of this rigid-body.

RoundShape
A shape with rounded borders.

SharedShape
The shape of a collider.

SolverContact
A contact seen by the constraints solver for computing forces.

SolverFlags
Flags affecting the behavior of the constraints solver for a given contact manifold.

TrackedContact
A single contact between two shape.

TriMesh
A triangle mesh.

Enums

CoefficientCombineRule
Rules used to combine two coefficients.

ColliderMassProps
The mass-properties of a collider.

ColliderPositionSync

ColliderType
The type of collider.

ContactEvent
Events occurring when two collision objects start or stop being in contact (or penetration).

FeatureId
An identifier of a feature of a convex polyhedron.

JointParams
An enum grouping all possible types of joints.

QueryPipelineMode
Indicates how the colliders position should be taken into account when updating the query pipeline.

RigidBodyPositionSync

RigidBodyType
The status of a body, governing the way it is affected by external forces.

SegmentPointLocation
Logical description of the location of a point on a triangle.

ShapeType
Enum representing the type of a shape.

SpringModel
The spring-like model used for constraints resolution.

TrianglePointLocation
Description of the location of a point on a triangle.

TypedShape
Enum representing the shape with its actual type

Constants

DEFAULT_EPSILON
The default tolerance used for geometric operations.

DIM
The dimension of the space.

MAX_MANIFOLD_POINTS
Max number of pairs of contact points from the same contact manifold that can be solved as part of a single contact constraint.

SIMD_LAST_INDEX
SIMD_WIDTH - 1

SIMD_WIDTH
The number of lanes of a SIMD number.

Traits

EventHandler
Trait implemented by structures responsible for handling events generated by the physics engine.

PhysicsHooks
User-defined functions called by the physics engines during one timestep in order to customize its behavior.

PhysicsHooksWithQuery

PolygonalFeatureMap
Trait implemented by convex shapes with features with polyhedral approximations.

Shape
Trait implemented by shapes usable by Rapier.

SimdCompositeShape
Trait implemented by shapes composed of multiple simpler shapes.

SupportMap
Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Definitions

AABB
An axis-aligned bounding box.

AngVector
The angular vector type.

AngularInertia
The angular inertia of a rigid body.

Ball
A ball shape.

BodyStatus	Deprecated
The type of a body, governing the way it is affected by external forces.

Capsule
A capsule shape.

ColliderGraphIndex
Index of a node of the interaction graph.

ColliderShape
The shape of a collider.

Contact
A contact between two colliders.

ContactManifold
A contact manifold between two colliders.

CrossMatrix
A matrix that represent the cross product with a given vector.

Cuboid
A cuboid shape.

DMatrix
A dynamically sized column-major matrix.

DVector
A dynamically sized column vector.

Dim
The dimension of the ambient space.

HeightField
A heightfield shape.

Isometry
The transformation matrix type.

Matrix
The matrix type.

NoUserData

Orientation
The orientation type.

Point
The point type.

PointProjection
The the projection of a point on a collider.

PrincipalAngularInertia
The principal angular inertia of a rigid body.

QueryPipelineColliderComponentsQuery

Ray
A ray that can be cast against colliders.

RayIntersection
The intersection between a ray and a collider.

Real
The scalar type used throughout this crate.

RigidBodyGraphIndex
Index of a node of the interaction graph.

Rotation
The rotation matrix type.

RoundConvexPolygon
A convex polygon dilated by a sphere (so it has round corners).

RoundCuboid
A cuboid dilated by a sphere (so it has round corners).

RoundTriangle
A triangle dilated by a sphere (so it has round corners).

SdpMatrix
A 2D symmetric-definite-positive matrix.

Segment
A segment shape.

SimdBool
A SIMD bool with SIMD_WIDTH lanes.

SimdReal
A SIMD float with SIMD_WIDTH lanes.

SpacialVector
A vector with a dimension equal to the maximum number of degrees of freedom of a rigid body.

TOI
The the time of impact between two shapes.

TemporaryInteractionIndex
Temporary index to and edge of the interaction graph.

Translation
The translation type.

Triangle
A triangle shape.

UnitVector
The unit vector type.

Vector
The vector type.