Module bevy_rapier2d::prelude::nalgebra

nalgebra

nalgebra is a linear algebra library written for Rust targeting:

General-purpose linear algebra (still lacks a lot of features…)

Real-time computer graphics.

Real-time computer physics.

Using nalgebra

You will need the last stable build of the rust compiler and the official package manager: cargo.

Simply add the following to your Cargo.toml file:

[dependencies]
// TODO: replace the * by the latest version.
nalgebra = "*"

Most useful functionalities of nalgebra are grouped in the root module nalgebra::.

However, the recommended way to use nalgebra is to import types and traits explicitly, and call free-functions using the na:: prefix:

#[macro_use]
extern crate approx; // For the macro relative_eq!
extern crate nalgebra as na;
use na::{Vector3, Rotation3};

fn main() {
    let axis  = Vector3::x_axis();
    let angle = 1.57;
    let b     = Rotation3::from_axis_angle(&axis, angle);

    relative_eq!(b.axis().unwrap(), axis);
    relative_eq!(b.angle(), angle);
}

Features

nalgebra is meant to be a general-purpose, low-dimensional, linear algebra library, with an optimized set of tools for computer graphics and physics. Those features include:

A single parametrizable type Matrix for vectors, (square or rectangular) matrices, and slices with dimensions known either at compile-time (using type-level integers) or at runtime.

Matrices and vectors with compile-time sizes are statically allocated while dynamic ones are allocated on the heap.

Convenient aliases for low-dimensional matrices and vectors: Vector1 to Vector6 and Matrix1x1 to Matrix6x6, including rectangular matrices like Matrix2x5.

Points sizes known at compile time, and convenience aliases: Point1 to Point6.

Translation (seen as a transformation that composes by multiplication): Translation2, Translation3.

Rotation matrices: Rotation2, Rotation3.

Quaternions: Quaternion, UnitQuaternion (for 3D rotation).

Unit complex numbers can be used for 2D rotation: UnitComplex.

Algebraic entities with a norm equal to one: Unit<T>, e.g., Unit<Vector3<f32>>.

Isometries (translation ⨯ rotation): Isometry2, Isometry3

Similarity transformations (translation ⨯ rotation ⨯ uniform scale): Similarity2, Similarity3.

Affine transformations stored as a homogeneous matrix: Affine2, Affine3.

Projective (i.e. invertible) transformations stored as a homogeneous matrix: Projective2, Projective3.

General transformations that does not have to be invertible, stored as a homogeneous matrix: Transform2, Transform3.

3D projections for computer graphics: Perspective3, Orthographic3.

Matrix factorizations: Cholesky, QR, LU, FullPivLU, SVD, Schur, Hessenberg, SymmetricEigen.

Insertion and removal of rows of columns of a matrix.

Modules

base
[Reexported at the root of this crate.] Data structures for vector and matrix computations.

geometry
[Reexported at the root of this crate.] Data structures for points and usual transformations (rotations, isometries, etc.)

linalg
[Reexported at the root of this crate.] Factorization of real matrices.

Macros

dmatrix
Construct a dynamic matrix directly from data.

dvector
Construct a dynamic column vector directly from data.

matrix
Construct a fixed-size matrix directly from data.

point
Construct a fixed-size point directly from data.

unimplemented_or_uninitialized_generic
When "no_unsound_assume_init" is enabled, expands to unimplemented!() instead of new_uninitialized_generic().assume_init(). Intended as a placeholder, each callsite should be refactored to use uninitialized memory soundly

vector
Construct a fixed-size column vector directly from data.

Structs

ArrayStorage
A array-based statically sized matrix data storage.

Bidiagonal
The bidiagonalization of a general matrix.

Cholesky
The Cholesky decomposition of a symmetric-definite-positive matrix.

ColPivQR
The QR decomposition (with column pivoting) of a general matrix.

Complex
A complex number in Cartesian form.

Const

DefaultAllocator
An allocator based on GenericArray and VecStorage for statically-sized and dynamically-sized matrices respectively.

DualQuaternion
A dual quaternion.

Dynamic
Dim of dynamically-sized algebraic entities.

EuclideanNorm
Euclidean norm.

FullPivLU
LU decomposition with full row and column pivoting.

Hessenberg
Hessenberg decomposition of a general matrix.

Isometry
A direct isometry, i.e., a rotation followed by a translation (aka. a rigid-body motion).

LU
LU decomposition with partial (row) pivoting.

LpNorm
Lp norm.

Matrix
The most generic column-major matrix (and vector) type.

Orthographic3
A 3D orthographic projection stored as a homogeneous 4x4 matrix.

PermutationSequence
A sequence of row or column permutations.

Perspective3
A 3D perspective projection stored as a homogeneous 4x4 matrix.

Point
A point in an euclidean space.

QR
The QR decomposition of a general matrix.

Quaternion
A quaternion. See the type alias UnitQuaternion = Unit<Quaternion> for a quaternion that may be used as a rotation.

Reflection
A reflection wrt. a plane.

Rotation
A rotation matrix.

SVD
Singular Value Decomposition of a general matrix.

Schur
Schur decomposition of a square matrix.

Similarity
A similarity, i.e., an uniform scaling, followed by a rotation, followed by a translation.

SliceStorage
A matrix data storage for a matrix slice. Only contains an internal reference to another matrix data storage.

SliceStorageMut
A mutable matrix data storage for mutable matrix slice. Only contains an internal mutable reference to another matrix data storage.

SymmetricEigen
Eigendecomposition of a symmetric matrix.

SymmetricTridiagonal
Tridiagonalization of a symmetric matrix.

Transform
A transformation matrix in homogeneous coordinates.

Translation
A translation.

UDU
UDU factorization.

UniformNorm
L-infinite norm aka. Chebytchev norm aka. uniform norm aka. suppremum norm.

Unit
A wrapper that ensures the underlying algebraic entity has a unit norm.

VecStorage
A Vec-based matrix data storage. It may be dynamically-sized.

Enums

TAffine
Tag representing an affine Transform. Its bottom-row is equal to (0, 0 ... 0, 1).

TGeneral
Tag representing the most general (not necessarily inversible) Transform type.

TProjective
Tag representing the most general inversible Transform type.

Traits

AbstractRotation
Trait implemented by rotations that can be used inside of an Isometry or Similarity.

ClosedAdd
Trait alias for Add and AddAssign with result of type Self.

ClosedDiv
Trait alias for Div and DivAssign with result of type Self.

ClosedMul
Trait alias for Mul and MulAssign with result of type Self.

ClosedSub
Trait alias for Sub and SubAssign with result of type Self.

ComplexField
Trait shared by all complex fields and its subfields (like real numbers).

Dim
Trait implemented by any type that can be used as a dimension. This includes type-level integers and Dynamic (for dimensions not known at compile-time).

DimAdd

DimDiv

DimMax

DimMin

DimMul

DimName
Trait implemented exclusively by type-level integers.

DimNameAdd

DimNameDiv

DimNameMax

DimNameMin

DimNameMul

DimNameSub

DimSub

Field
Trait implemented by fields, i.e., complex numbers and floats.

IsDynamic
Trait implemented by Dynamic.

IsNotStaticOne
Trait implemented by Dynamic and type-level integers different from U1.

Norm
A trait for abstract matrix norms.

Normed
Trait implemented by entities scan be be normalized and put in an Unit struct.

RealField
Trait shared by all reals.

Scalar
The basic scalar type for all structures of nalgebra.

SimdBool
Lane-wise generalization of bool for SIMD booleans.

SimdComplexField
Lane-wise generalisation of ComplexField for SIMD complex fields.

SimdPartialOrd
Lane-wise generalization of the standard PartialOrd for SIMD values.

SimdRealField
Lanewise generalization of RealField for SIMD reals.

SimdValue
Base trait for every SIMD types.

SliceRange
A range with a size that may be known at compile-time.

SubTCategoryOf
Indicates that Self is a more specific Transform category than Other.

SuperTCategoryOf
Indicates that Self is a more general Transform category than Other.

TCategory
Trait implemented by phantom types identifying the projective transformation type.

TCategoryMul
Traits that gives the Transform category that is compatible with the result of the multiplication of transformations with categories Self and Other.

ToConst

ToTypenum

Functions

abs	Deprecated
The absolute value of a.

center
The center of two points.

clamp
Returns a reference to the input value clamped to the interval [min, max].

convert
Converts an object from one type to an equivalent or more general one.

convert_ref
Converts an object from one type to an equivalent or more general one.

convert_ref_unchecked
Use with care! Same as try_convert but without any property checks.

convert_unchecked
Use with care! Same as try_convert but without any property checks.

distance
The distance between two points.

distance_squared
The squared distance between two points.

inf	Deprecated
Returns the infimum of a and b.

inf_sup	Deprecated
Returns simultaneously the infimum and supremum of a and b.

is_convertible
Indicates if try_convert will succeed without actually performing the conversion.

max
Same as cmp::max.

min
Same as cmp::min.

one
Gets the multiplicative identity element.

partial_clamp
Clamp value between min and max. Returns None if value is not comparable to min or max.

partial_cmp
Compare a and b using a partial ordering relation.

partial_ge
Returns true iff a and b are comparable and a >= b.

partial_gt
Returns true iff a and b are comparable and a > b.

partial_le
Returns true iff a and b are comparable and a <= b.

partial_lt
Returns true iff a and b are comparable and a < b.

partial_max
Return the maximum of a and b if they are comparable.

partial_min
Return the minimum of a and b if they are comparable.

partial_sort2
Sorts two values in increasing order using a partial ordering.

sup	Deprecated
Returns the supremum of a and b.

try_convert
Attempts to convert an object to a more specific one.

try_convert_ref
Attempts to convert an object to a more specific one.

try_invert_to
Performs a LU decomposition to overwrite out with the inverse of matrix.

wilkinson_shift
Computes the wilkinson shift, i.e., the 2x2 symmetric matrix eigenvalue to its tailing component tnn.

wrap
Wraps val into the range [min, max] using modular arithmetics.

zero
Gets the additive identity element.

Type Definitions

Affine2
A 2D affine transformation. Stored as a homogeneous 3x3 matrix.

Affine3
A 3D affine transformation. Stored as a homogeneous 4x4 matrix.

DMatrix
A dynamically sized column-major matrix.

DMatrixSlice
A column-major matrix slice dynamic numbers of rows and columns.

DMatrixSliceMut
A column-major matrix slice dynamic numbers of rows and columns.

DVector
A dynamically sized column vector.

DVectorSlice
A column vector slice dynamic numbers of rows and columns.

DVectorSliceMut
A column vector slice dynamic numbers of rows and columns.

DimDiff

DimMaximum

DimMinimum

DimNameDiff

DimNameMaximum

DimNameMinimum

DimNameProd

DimNameQuot

DimNameSum

DimProd

DimQuot

DimSum

Isometry2
A 2-dimensional direct isometry using a unit complex number for its rotational part.

Isometry3
A 3-dimensional direct isometry using a unit quaternion for its rotational part.

IsometryMatrix2
A 2-dimensional direct isometry using a rotation matrix for its rotational part.

IsometryMatrix3
A 3-dimensional direct isometry using a rotation matrix for its rotational part.

Matrix1
A stack-allocated, column-major, 1x1 square matrix.

Matrix1x2
A stack-allocated, column-major, 1x2 matrix.

Matrix1x3
A stack-allocated, column-major, 1x3 matrix.

Matrix1x4
A stack-allocated, column-major, 1x4 matrix.

Matrix1x5
A stack-allocated, column-major, 1x5 matrix.

Matrix1x6
A stack-allocated, column-major, 1x6 matrix.

Matrix1xX
A heap-allocated, row-major, matrix with 1 rows and a dynamic number of columns.

Matrix2
A stack-allocated, column-major, 2x2 square matrix.

Matrix2x1
A stack-allocated, column-major, 2x1 matrix.

Matrix2x3
A stack-allocated, column-major, 2x3 matrix.

Matrix2x4
A stack-allocated, column-major, 2x4 matrix.

Matrix2x5
A stack-allocated, column-major, 2x5 matrix.

Matrix2x6
A stack-allocated, column-major, 2x6 matrix.

Matrix2xX
A heap-allocated, row-major, matrix with 2 rows and a dynamic number of columns.

Matrix3
A stack-allocated, column-major, 3x3 square matrix.

Matrix3x1
A stack-allocated, column-major, 3x1 matrix.

Matrix3x2
A stack-allocated, column-major, 3x2 matrix.

Matrix3x4
A stack-allocated, column-major, 3x4 matrix.

Matrix3x5
A stack-allocated, column-major, 3x5 matrix.

Matrix3x6
A stack-allocated, column-major, 3x6 matrix.

Matrix3xX
A heap-allocated, row-major, matrix with 3 rows and a dynamic number of columns.

Matrix4
A stack-allocated, column-major, 4x4 square matrix.

Matrix4x1
A stack-allocated, column-major, 4x1 matrix.

Matrix4x2
A stack-allocated, column-major, 4x2 matrix.

Matrix4x3
A stack-allocated, column-major, 4x3 matrix.

Matrix4x5
A stack-allocated, column-major, 4x5 matrix.

Matrix4x6
A stack-allocated, column-major, 4x6 matrix.

Matrix4xX
A heap-allocated, row-major, matrix with 4 rows and a dynamic number of columns.

Matrix5
A stack-allocated, column-major, 5x5 square matrix.

Matrix5x1
A stack-allocated, column-major, 5x1 matrix.

Matrix5x2
A stack-allocated, column-major, 5x2 matrix.

Matrix5x3
A stack-allocated, column-major, 5x3 matrix.

Matrix5x4
A stack-allocated, column-major, 5x4 matrix.

Matrix5x6
A stack-allocated, column-major, 5x6 matrix.

Matrix5xX
A heap-allocated, row-major, matrix with 5 rows and a dynamic number of columns.

Matrix6
A stack-allocated, column-major, 6x6 square matrix.

Matrix6x1
A stack-allocated, column-major, 6x1 matrix.

Matrix6x2
A stack-allocated, column-major, 6x2 matrix.

Matrix6x3
A stack-allocated, column-major, 6x3 matrix.

Matrix6x4
A stack-allocated, column-major, 6x4 matrix.

Matrix6x5
A stack-allocated, column-major, 6x5 matrix.

Matrix6xX
A heap-allocated, row-major, matrix with 6 rows and a dynamic number of columns.

MatrixCross
The type of the result of a matrix cross product.

MatrixMN	Deprecated
An owned matrix column-major matrix with R rows and C columns.

MatrixN	Deprecated
An owned matrix column-major matrix with D columns.

MatrixSlice
A matrix slice.

MatrixSlice1
A column-major 1x1 matrix slice.

MatrixSlice1x2
A column-major 1x2 matrix slice.

MatrixSlice1x3
A column-major 1x3 matrix slice.

MatrixSlice1x4
A column-major 1x4 matrix slice.

MatrixSlice1x5
A column-major 1x5 matrix slice.

MatrixSlice1x6
A column-major 1x6 matrix slice.

MatrixSlice1xX
A column-major matrix slice with 1 row and a number of columns chosen at runtime.

MatrixSlice2
A column-major 2x2 matrix slice.

MatrixSlice2x1
A column-major 2x1 matrix slice.

MatrixSlice2x3
A column-major 2x3 matrix slice.

MatrixSlice2x4
A column-major 2x4 matrix slice.

MatrixSlice2x5
A column-major 2x5 matrix slice.

MatrixSlice2x6
A column-major 2x6 matrix slice.

MatrixSlice2xX
A column-major matrix slice with 2 rows and a number of columns chosen at runtime.

MatrixSlice3
A column-major 3x3 matrix slice.

MatrixSlice3x1
A column-major 3x1 matrix slice.

MatrixSlice3x2
A column-major 3x2 matrix slice.

MatrixSlice3x4
A column-major 3x4 matrix slice.

MatrixSlice3x5
A column-major 3x5 matrix slice.

MatrixSlice3x6
A column-major 3x6 matrix slice.

MatrixSlice3xX
A column-major matrix slice with 3 rows and a number of columns chosen at runtime.

MatrixSlice4
A column-major 4x4 matrix slice.

MatrixSlice4x1
A column-major 4x1 matrix slice.

MatrixSlice4x2
A column-major 4x2 matrix slice.

MatrixSlice4x3
A column-major 4x3 matrix slice.

MatrixSlice4x5
A column-major 4x5 matrix slice.

MatrixSlice4x6
A column-major 4x6 matrix slice.

MatrixSlice4xX
A column-major matrix slice with 4 rows and a number of columns chosen at runtime.

MatrixSlice5
A column-major 5x5 matrix slice.

MatrixSlice5x1
A column-major 5x1 matrix slice.

MatrixSlice5x2
A column-major 5x2 matrix slice.

MatrixSlice5x3
A column-major 5x3 matrix slice.

MatrixSlice5x4
A column-major 5x4 matrix slice.

MatrixSlice5x6
A column-major 5x6 matrix slice.

MatrixSlice5xX
A column-major matrix slice with 5 rows and a number of columns chosen at runtime.

MatrixSlice6
A column-major 6x6 matrix slice.

MatrixSlice6x1
A column-major 6x1 matrix slice.

MatrixSlice6x2
A column-major 6x2 matrix slice.

MatrixSlice6x3
A column-major 6x3 matrix slice.

MatrixSlice6x4
A column-major 6x4 matrix slice.

MatrixSlice6x5
A column-major 6x5 matrix slice.

MatrixSlice6xX
A column-major matrix slice with 6 rows and a number of columns chosen at runtime.

MatrixSliceMut
A mutable matrix slice.

MatrixSliceMut1
A column-major 1x1 matrix slice.

MatrixSliceMut1x2
A column-major 1x2 matrix slice.

MatrixSliceMut1x3
A column-major 1x3 matrix slice.

MatrixSliceMut1x4
A column-major 1x4 matrix slice.

MatrixSliceMut1x5
A column-major 1x5 matrix slice.

MatrixSliceMut1x6
A column-major 1x6 matrix slice.

MatrixSliceMut1xX
A column-major matrix slice with 1 row and a number of columns chosen at runtime.

MatrixSliceMut2
A column-major 2x2 matrix slice.

MatrixSliceMut2x1
A column-major 2x1 matrix slice.

MatrixSliceMut2x3
A column-major 2x3 matrix slice.

MatrixSliceMut2x4
A column-major 2x4 matrix slice.

MatrixSliceMut2x5
A column-major 2x5 matrix slice.

MatrixSliceMut2x6
A column-major 2x6 matrix slice.

MatrixSliceMut2xX
A column-major matrix slice with 2 rows and a number of columns chosen at runtime.

MatrixSliceMut3
A column-major 3x3 matrix slice.

MatrixSliceMut3x1
A column-major 3x1 matrix slice.

MatrixSliceMut3x2
A column-major 3x2 matrix slice.

MatrixSliceMut3x4
A column-major 3x4 matrix slice.

MatrixSliceMut3x5
A column-major 3x5 matrix slice.

MatrixSliceMut3x6
A column-major 3x6 matrix slice.

MatrixSliceMut3xX
A column-major matrix slice with 3 rows and a number of columns chosen at runtime.

MatrixSliceMut4
A column-major 4x4 matrix slice.

MatrixSliceMut4x1
A column-major 4x1 matrix slice.

MatrixSliceMut4x2
A column-major 4x2 matrix slice.

MatrixSliceMut4x3
A column-major 4x3 matrix slice.

MatrixSliceMut4x5
A column-major 4x5 matrix slice.

MatrixSliceMut4x6
A column-major 4x6 matrix slice.

MatrixSliceMut4xX
A column-major matrix slice with 4 rows and a number of columns chosen at runtime.

MatrixSliceMut5
A column-major 5x5 matrix slice.

MatrixSliceMut5x1
A column-major 5x1 matrix slice.

MatrixSliceMut5x2
A column-major 5x2 matrix slice.

MatrixSliceMut5x3
A column-major 5x3 matrix slice.

MatrixSliceMut5x4
A column-major 5x4 matrix slice.

MatrixSliceMut5x6
A column-major 5x6 matrix slice.

MatrixSliceMut5xX
A column-major matrix slice with 5 rows and a number of columns chosen at runtime.

MatrixSliceMut6
A column-major 6x6 matrix slice.

MatrixSliceMut6x1
A column-major 6x1 matrix slice.

MatrixSliceMut6x2
A column-major 6x2 matrix slice.

MatrixSliceMut6x3
A column-major 6x3 matrix slice.

MatrixSliceMut6x4
A column-major 6x4 matrix slice.

MatrixSliceMut6x5
A column-major 6x5 matrix slice.

MatrixSliceMut6xX
A column-major matrix slice with 6 rows and a number of columns chosen at runtime.

MatrixSliceMutMN
A column-major matrix slice with R rows and C columns.

MatrixSliceMutN
A column-major matrix slice with D rows and columns.

MatrixSliceMutXx1
A column-major matrix slice with a number of rows chosen at runtime and 1 column.

MatrixSliceMutXx2
A column-major matrix slice with a number of rows chosen at runtime and 2 columns.

MatrixSliceMutXx3
A column-major matrix slice with a number of rows chosen at runtime and 3 columns.

MatrixSliceMutXx4
A column-major matrix slice with a number of rows chosen at runtime and 4 columns.

MatrixSliceMutXx5
A column-major matrix slice with a number of rows chosen at runtime and 5 columns.

MatrixSliceMutXx6
A column-major matrix slice with a number of rows chosen at runtime and 6 columns.

MatrixSliceXx1
A column-major matrix slice with a number of rows chosen at runtime and 1 column.

MatrixSliceXx2
A column-major matrix slice with a number of rows chosen at runtime and 2 columns.

MatrixSliceXx3
A column-major matrix slice with a number of rows chosen at runtime and 3 columns.

MatrixSliceXx4
A column-major matrix slice with a number of rows chosen at runtime and 4 columns.

MatrixSliceXx5
A column-major matrix slice with a number of rows chosen at runtime and 5 columns.

MatrixSliceXx6
A column-major matrix slice with a number of rows chosen at runtime and 6 columns.

MatrixSum
The type of the result of a matrix sum.

MatrixVec	Deprecated
Renamed to VecStorage.

MatrixXx1
A heap-allocated, column-major, matrix with a dynamic number of rows and 1 columns.

MatrixXx2
A heap-allocated, column-major, matrix with a dynamic number of rows and 2 columns.

MatrixXx3
A heap-allocated, column-major, matrix with a dynamic number of rows and 3 columns.

MatrixXx4
A heap-allocated, column-major, matrix with a dynamic number of rows and 4 columns.

MatrixXx5
A heap-allocated, column-major, matrix with a dynamic number of rows and 5 columns.

MatrixXx6
A heap-allocated, column-major, matrix with a dynamic number of rows and 6 columns.

OMatrix
An owned matrix column-major matrix with R rows and C columns.

OVector
An owned D-dimensional column vector.

Point1
A statically sized 1-dimensional column point.

Point2
A statically sized 2-dimensional column point.

Point3
A statically sized 3-dimensional column point.

Point4
A statically sized 4-dimensional column point.

Point5
A statically sized 5-dimensional column point.

Point6
A statically sized 6-dimensional column point.

Projective2
An invertible 2D general transformation. Stored as a homogeneous 3x3 matrix.

Projective3
An invertible 3D general transformation. Stored as a homogeneous 4x4 matrix.

Rotation2
A 2-dimensional rotation matrix.

Rotation3
A 3-dimensional rotation matrix.

RowDVector
A dynamically sized row vector.

RowOVector
An owned D-dimensional row vector.

RowSVector
A statically sized D-dimensional row vector.

RowVector
A matrix with one row and D columns .

RowVector1
A stack-allocated, 1-dimensional row vector.

RowVector2
A stack-allocated, 2-dimensional row vector.

RowVector3
A stack-allocated, 3-dimensional row vector.

RowVector4
A stack-allocated, 4-dimensional row vector.

RowVector5
A stack-allocated, 5-dimensional row vector.

RowVector6
A stack-allocated, 6-dimensional row vector.

SMatrix
A statically sized column-major matrix with R rows and C columns.

SMatrixSlice
A column-major matrix slice with dimensions known at compile-time.

SMatrixSliceMut
A column-major matrix slice with dimensions known at compile-time.

SVector
A statically sized D-dimensional column vector.

SVectorSlice
A column vector slice with dimensions known at compile-time.

SVectorSliceMut
A column vector slice with dimensions known at compile-time.

Similarity2
A 2-dimensional similarity.

Similarity3
A 3-dimensional similarity.

SimilarityMatrix2
A 2-dimensional similarity using a rotation matrix for its rotation part.

SimilarityMatrix3
A 3-dimensional similarity using a rotation matrix for its rotation part.

SquareMatrix
A square matrix.

Transform2
A 2D general transformation that may not be invertible. Stored as a homogeneous 3x3 matrix.

Transform3
A 3D general transformation that may not be inversible. Stored as a homogeneous 4x4 matrix.

Translation1
A 1-dimensional translation.

Translation2
A 2-dimensional translation.

Translation3
A 3-dimensional translation.

Translation4
A 4-dimensional translation.

Translation5
A 5-dimensional translation.

Translation6
A 6-dimensional translation.

U0

U1

U2

U3

U4

U5

U6

U7

U8

U9

U10

U11

U12

U13

U14

U15

U16

U17

U18

U19

U20

U21

U22

U23

U24

U25

U26

U27

U28

U29

U30

U31

U32

U33

U34

U35

U36

U37

U38

U39

U40

U41

U42

U43

U44

U45

U46

U47

U48

U49

U50

U51

U52

U53

U54

U55

U56

U57

U58

U59

U60

U61

U62

U63

U64

U65

U66

U67

U68

U69

U70

U71

U72

U73

U74

U75

U76

U77

U78

U79

U80

U81

U82

U83

U84

U85

U86

U87

U88

U89

U90

U91

U92

U93

U94

U95

U96

U97

U98

U99

U100

U101

U102

U103

U104

U105

U106

U107

U108

U109

U110

U111

U112

U113

U114

U115

U116

U117

U118

U119

U120

U121

U122

U123

U124

U125

U126

U127

UnitComplex
A 2D rotation represented as a complex number with magnitude 1.

UnitDualQuaternion
A unit quaternions. May be used to represent a rotation followed by a translation.

UnitQuaternion
A unit quaternions. May be used to represent a rotation.

UnitVector1
A stack-allocated, 1-dimensional unit vector.

UnitVector2
A stack-allocated, 2-dimensional unit vector.

UnitVector3
A stack-allocated, 3-dimensional unit vector.

UnitVector4
A stack-allocated, 4-dimensional unit vector.

UnitVector5
A stack-allocated, 5-dimensional unit vector.

UnitVector6
A stack-allocated, 6-dimensional unit vector.

Vector
A matrix with one column and D rows.

Vector1
A stack-allocated, 1-dimensional column vector.

Vector2
A stack-allocated, 2-dimensional column vector.

Vector3
A stack-allocated, 3-dimensional column vector.

Vector4
A stack-allocated, 4-dimensional column vector.

Vector5
A stack-allocated, 5-dimensional column vector.

Vector6
A stack-allocated, 6-dimensional column vector.

VectorN	Deprecated
An owned matrix column-major matrix with R rows and C columns.

VectorSlice
A column vector slice with dimensions known at compile-time.

VectorSlice1
A 1D column vector slice.

VectorSlice2
A 2D column vector slice.

VectorSlice3
A 3D column vector slice.

VectorSlice4
A 4D column vector slice.

VectorSlice5
A 5D column vector slice.

VectorSlice6
A 6D column vector slice.

VectorSliceMut
A column vector slice with dimensions known at compile-time.

VectorSliceMut1
A 1D column vector slice.

VectorSliceMut2
A 2D column vector slice.

VectorSliceMut3
A 3D column vector slice.

VectorSliceMut4
A 4D column vector slice.

VectorSliceMut5
A 5D column vector slice.

VectorSliceMut6
A 6D column vector slice.

VectorSum
The type of the result of a matrix sum.