Trait bevy_rapier2d::prelude::nalgebra::base::storage::ContiguousStorage

pub unsafe trait ContiguousStorage<T, R, C = Const<1_usize>>: Storage<T, R, C> where
    C: Dim,
    T: Scalar,
    R: Dim, { }

A matrix storage that is stored contiguously in memory.

The storage requirement means that for any value of i in [0, nrows * ncols - 1], the value .get_unchecked_linear returns one of the matrix component. This trait is unsafe because failing to comply to this may cause Undefined Behaviors.