Type Definition bevy_rapier2d::prelude::nalgebra::base::MatrixXx5

type MatrixXx5<T> = Matrix<T, Dynamic, Const<{_: usize}>, VecStorage<T, Dynamic, Const<{_: usize}>>>;

A heap-allocated, column-major, matrix with a dynamic number of rows and 5 columns.

Because this is an alias, not all its methods are listed here. See the Matrix type too.