Trait bevy_rapier2d::prelude::nalgebra::AbstractRotation

pub trait AbstractRotation<T, const D: usize>: Clone + PartialEq<Self> + ClosedMul<Self> where
    T: Scalar, {
    pub fn identity() -> Self;

    pub fn inverse(&self) -> Self;

    pub fn inverse_mut(&mut self);

    pub fn transform_vector(
        &self, 
        v: &Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>
    ) -> Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>;

    pub fn transform_point(&self, p: &Point<T, D>) -> Point<T, D>;

    pub fn inverse_transform_vector(
        &self, 
        v: &Matrix<T, Const<D>, Const<1_usize>, <DefaultAllocator as Allocator<T, Const<D>, Const<1_usize>>>::Buffer>
    ) -> Matrix<T, Const<D>, Const<1_usize>, <DefaultAllocator as Allocator<T, Const<D>, Const<1_usize>>>::Buffer>;

    pub fn inverse_transform_point(&self, p: &Point<T, D>) -> Point<T, D>;

    pub fn inverse_transform_unit_vector(
        &self, 
        v: &Unit<Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>>
    ) -> Unit<Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>> { ... }
}

Trait implemented by rotations that can be used inside of an Isometry or Similarity.

Required methods

pub fn identity() -> Self

The rotation identity.

pub fn inverse(&self) -> Self

The rotation inverse.

pub fn inverse_mut(&mut self)

Change self to its inverse.

pub fn transform_vector(
&self,
v: &Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>
) -> Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>

Apply the rotation to the given vector.

pub fn transform_point(&self, p: &Point<T, D>) -> Point<T, D>

Apply the rotation to the given point.

pub fn inverse_transform_vector(
&self,
v: &Matrix<T, Const<D>, Const<1_usize>, <DefaultAllocator as Allocator<T, Const<D>, Const<1_usize>>>::Buffer>
) -> Matrix<T, Const<D>, Const<1_usize>, <DefaultAllocator as Allocator<T, Const<D>, Const<1_usize>>>::Buffer>

Apply the inverse rotation to the given vector.

pub fn inverse_transform_point(&self, p: &Point<T, D>) -> Point<T, D>

Apply the inverse rotation to the given point.

Loading content...
Provided methods

pub fn inverse_transform_unit_vector(
&self,
v: &Unit<Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>>
) -> Unit<Matrix<T, Const<D>, Const<1_usize>, ArrayStorage<T, D, 1_usize>>>

Apply the inverse rotation to the given unit vector.

Loading content...