Trait bevy_rapier2d::prelude::nalgebra::SliceRange

pub trait SliceRange<D> where
    D: Dim, {
    type Size: Dim;
    pub fn begin(&self, shape: D) -> usize;

    pub fn end(&self, shape: D) -> usize;

    pub fn size(&self, shape: D) -> Self::Size;
}

A range with a size that may be known at compile-time.

This may be:

A single usize index, e.g., 4

A left-open range std::ops::RangeTo, e.g., .. 4

A right-open range std::ops::RangeFrom, e.g., 4 ..

A full range std::ops::RangeFull, e.g., ..

Associated Types

type Size: Dim

Type of the range size. May be a type-level integer.

Loading content...
Required methods

pub fn begin(&self, shape: D) -> usize

The start index of the range.

pub fn end(&self, shape: D) -> usize

The index immediately after the last index inside of the range.

pub fn size(&self, shape: D) -> Self::Size

The number of elements of the range, i.e., self.end - self.begin.

Loading content...