Trait bevy_rapier2d::prelude::nalgebra::TCategory

pub trait TCategory: PartialEq<Self> + Send + Copy + Any + Debug {
    pub fn check_homogeneous_invariants<T, D>(
        mat: &Matrix<T, D, D, <DefaultAllocator as Allocator<T, D, D>>::Buffer>
    ) -> bool
    where
        T: RealField,
        D: DimName,
        <T as AbsDiffEq<T>>::Epsilon: Copy,
        DefaultAllocator: Allocator<T, D, D>;

    pub fn has_normalizer() -> bool { ... }
}

Trait implemented by phantom types identifying the projective transformation type.

NOTE: this trait is not intended to be implemented outside of the nalgebra crate.

Required methods

pub fn check_homogeneous_invariants<T, D>(
mat: &Matrix<T, D, D, <DefaultAllocator as Allocator<T, D, D>>::Buffer>
) -> bool where
T: RealField,
D: DimName,
<T as AbsDiffEq<T>>::Epsilon: Copy,
DefaultAllocator: Allocator<T, D, D>,

Checks that the given matrix is a valid homogeneous representation of an element of the category Self.

Loading content...
Provided methods

pub fn has_normalizer() -> bool

Indicates whether a Transform with the category Self has a bottom-row different from 0 0 .. 1.

Loading content...