Struct bevy_rapier2d::prelude::PrismaticJoint

pub struct PrismaticJoint {
    pub local_anchor1: Point<f32, 2_usize>,
    pub local_anchor2: Point<f32, 2_usize>,
    pub impulse: Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>,
    pub limits_enabled: bool,
    pub limits: [f32; 2],
    pub limits_impulse: f32,
    pub motor_target_vel: f32,
    pub motor_target_pos: f32,
    pub motor_stiffness: f32,
    pub motor_damping: f32,
    pub motor_max_impulse: f32,
    pub motor_impulse: f32,
    pub motor_model: SpringModel,
    // some fields omitted
}

A joint that removes all relative motion between two bodies, except for the translations along one axis.

Fields
local_anchor1: Point<f32, 2_usize>

Where the prismatic joint is attached on the first body, expressed in the local space of the first attached body.

local_anchor2: Point<f32, 2_usize>

Where the prismatic joint is attached on the second body, expressed in the local space of the second attached body.

impulse: Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>

The impulse applied by this joint on the first body.

The impulse applied to the second body is given by -impulse.

limits_enabled: bool

Whether or not this joint should enforce translational limits along its axis.

limits: [f32; 2]

The min an max relative position of the attached bodies along this joint's axis.

limits_impulse: f32

The impulse applied by this joint on the first body to enforce the position limit along this joint's axis.

The impulse applied to the second body is given by -impulse.

motor_target_vel: f32

The target relative angular velocity the motor will attempt to reach.

motor_target_pos: f32

The target relative angle along the joint axis the motor will attempt to reach.

motor_stiffness: f32

The motor's stiffness. See the documentation of SpringModel for more information on this parameter.

motor_damping: f32

The motor's damping. See the documentation of SpringModel for more information on this parameter.

motor_max_impulse: f32

The maximal impulse the motor is able to deliver.

motor_impulse: f32

The angular impulse applied by the motor.

motor_model: SpringModel

The spring-like model used by the motor to reach the target velocity and .

Implementations

impl PrismaticJoint

pub fn new(
local_anchor1: Point<f32, 2_usize>,
local_axis1: Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>,
local_anchor2: Point<f32, 2_usize>,
local_axis2: Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>
) -> PrismaticJoint

Creates a new prismatic joint with the given point of applications and axis, all expressed in the local-space of the affected bodies.

pub fn local_axis1(
&self
) -> Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>

The local axis of this joint, expressed in the local-space of the first attached body.

pub fn local_axis2(
&self
) -> Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>

The local axis of this joint, expressed in the local-space of the second attached body.

pub fn supports_simd_constraints(&self) -> bool

Can a SIMD constraint be used for resolving this joint?

pub fn configure_motor_model(&mut self, model: SpringModel)

Set the spring-like model used by the motor to reach the desired target velocity and position.

pub fn configure_motor_velocity(&mut self, target_vel: f32, factor: f32)

Sets the target velocity this motor needs to reach.

pub fn configure_motor_position(
&mut self,
target_pos: f32,
stiffness: f32,
damping: f32
)

Sets the target position this motor needs to reach.

pub fn configure_motor(
&mut self,
target_pos: f32,
target_vel: f32,
stiffness: f32,
damping: f32
)

Configure both the target position and target velocity of the motor.