Trait bevy_rapier2d::prelude::SupportMap

pub trait SupportMap {
    pub fn local_support_point(
        &self, 
        dir: &Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>
    ) -> Point<f32, 2_usize>;

    pub fn local_support_point_toward(
        &self, 
        dir: &Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>
    ) -> Point<f32, 2_usize> { ... }

    pub fn support_point(
        &self, 
        transform: &Isometry<f32, Unit<Complex<f32>>, 2_usize>, 
        dir: &Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>
    ) -> Point<f32, 2_usize> { ... }

    pub fn support_point_toward(
        &self, 
        transform: &Isometry<f32, Unit<Complex<f32>>, 2_usize>, 
        dir: &Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>
    ) -> Point<f32, 2_usize> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required methods

pub fn local_support_point(
&self,
dir: &Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>
) -> Point<f32, 2_usize>

Loading content...
Provided methods

pub fn local_support_point_toward(
&self,
dir: &Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>
) -> Point<f32, 2_usize>

Same as self.local_support_point except that dir is normalized.

pub fn support_point(
&self,
transform: &Isometry<f32, Unit<Complex<f32>>, 2_usize>,
dir: &Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>
) -> Point<f32, 2_usize>

pub fn support_point_toward(
&self,
transform: &Isometry<f32, Unit<Complex<f32>>, 2_usize>,
dir: &Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>
) -> Point<f32, 2_usize>

Same as self.support_point except that dir is normalized.

Loading content...